Consolidate saltclass pillar unit tests into single module
